requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
dotenv==0.9.9
frozenlist==1.5.0
http_ece==1.2.1
idna==3.10
multidict==6.1.0
propcache==0.3.0
py-vapid==1.9.2
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2025.1
pywebpush==2.0.3
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
yarl==1.18.3
